Split template into common and specific
Figured out how I want to consistently do simulation of events and
rescanning. Template is split into common part to both and individual
parts which need to be attached to common part. In
SetSindarinSimulationFiles.py script the parameters put into the
template are saved so that the rescanning can be sure to use the same
parameters. Rescanning (not implemented) will then only overwrite
parameters which change in scenario (not e.g. final state, cuts, ...).